requirements.txt
azure-functions
firebase-admin